+++ crates/vm/levm/src/opcode_handlers/keccak.rs
@@ -1,7 +1,7 @@
 use crate::{
     errors::{ExceptionalHalt, OpcodeResult, VMError},
     gas_cost,
-    memory::{self, calculate_memory_size},
+    memory::calculate_memory_size,
     vm::VM,
 };
 use ethrex_common::utils::u256_from_big_endian;
@@ -17,6 +17,10 @@ impl<'a> VM<'a> {
         let size: usize = size
             .try_into()
             .map_err(|_| ExceptionalHalt::VeryLargeNumber)?;
+        let offset: usize = match offset.try_into() {
+            Ok(x) => x,
+            Err(_) => usize::MAX,
+        };
 
         let new_memory_size = calculate_memory_size(offset, size)?;
 
@@ -27,11 +31,7 @@ impl<'a> VM<'a> {
         )?)?;
 
         let mut hasher = Keccak256::new();
-        hasher.update(memory::load_range(
-            &mut current_call_frame.memory,
-            offset,
-            size,
-        )?);
+        hasher.update(current_call_frame.memory.load_range(offset, size)?);
         current_call_frame
             .stack
             .push(&[u256_from_big_endian(&hasher.finalize())])?;