+++ crates/vm/levm/src/opcode_handlers/push.rs
@@ -1,10 +1,8 @@
 use crate::{
-    call_frame::CallFrame,
     errors::{ExceptionalHalt, InternalError, OpcodeResult, VMError},
     gas_cost,
     vm::VM,
 };
-use ExceptionalHalt::OutOfBounds;
 use ethrex_common::{U256, types::Fork, utils::u256_from_big_endian_const};
 
 // Push Operations
@@ -16,9 +14,35 @@ impl<'a> VM<'a> {
         let current_call_frame = self.current_call_frame_mut()?;
         current_call_frame.increase_consumed_gas(gas_cost::PUSHN)?;
 
-        let read_n_bytes = read_bytcode_slice::<N>(current_call_frame)?;
+        let current_pc = current_call_frame.pc;
 
-        let value = u256_from_big_endian_const(read_n_bytes);
+        // Check to avoid multiple checks.
+        if current_pc.checked_add(N.wrapping_add(1)).is_none() {
+            Err(InternalError::Overflow)?;
+        }
+
+        let pc_offset = current_pc
+            // Add 1 to the PC because we don't want to include the
+            // Bytecode of the current instruction in the data we're about
+            // to read. We only want to read the data _NEXT_ to that
+            // bytecode
+            .wrapping_add(1);
+
+        let value = if let Some(slice) = current_call_frame
+            .bytecode
+            .get(pc_offset..pc_offset.wrapping_add(N))
+        {
+            u256_from_big_endian_const(
+                // SAFETY: If the get succeeded, we got N elements so the cast is safe.
+                #[expect(unsafe_code)]
+                unsafe {
+                    *slice.as_ptr().cast::<[u8; N]>()
+                },
+            )
+        } else {
+            U256::zero()
+        };
+
         current_call_frame.stack.push(&[value])?;
 
         // The n_bytes that you push to the stack + 1 for the next instruction
@@ -42,28 +66,5 @@ impl<'a> VM<'a> {
         current_call_frame.stack.push(&[U256::zero()])?;
 
         Ok(OpcodeResult::Continue { pc_increment: 1 })
-    }
-}
-
-// Like `read_bytcode_slice` but using a const generic and returning a fixed size array.
-fn read_bytcode_slice<const N: usize>(current_call_frame: &CallFrame) -> Result<[u8; N], VMError> {
-    let current_pc = current_call_frame.pc;
-    let pc_offset = current_pc
-        // Add 1 to the PC because we don't want to include the
-        // Bytecode of the current instruction in the data we're about
-        // to read. We only want to read the data _NEXT_ to that
-        // bytecode
-        .checked_add(1)
-        .ok_or(InternalError::Overflow)?;
-
-    if let Some(slice) = current_call_frame
-        .bytecode
-        .get(pc_offset..pc_offset.checked_add(N).ok_or(OutOfBounds)?)
-    {
-        Ok(slice
-            .try_into()
-            .map_err(|_| VMError::Internal(InternalError::TypeConversion))?)
-    } else {
-        Ok([0; N])
     }
 }