+++ crates/vm/levm/src/opcode_handlers/system.rs
@@ -3,7 +3,7 @@ use crate::{
     constants::{FAIL, INIT_CODE_MAX_SIZE, SUCCESS},
     errors::{ContextResult, ExceptionalHalt, InternalError, OpcodeResult, TxResult, VMError},
     gas_cost::{self, max_message_call_gas},
-    memory::{self, calculate_memory_size},
+    memory::calculate_memory_size,
     utils::{address_to_word, word_to_address, *},
     vm::VM,
 };
@@ -40,7 +40,13 @@ impl<'a> VM<'a> {
                 return_data_size,
             ] = *current_call_frame.stack.pop()?;
             let callee: Address = word_to_address(callee);
+            let args_start_offset = args_start_offset
+                .try_into()
+                .map_err(|_| ExceptionalHalt::VeryLargeNumber)?;
             let args_size = args_size
+                .try_into()
+                .map_err(|_| ExceptionalHalt::VeryLargeNumber)?;
+            let return_data_start_offset: usize = return_data_start_offset
                 .try_into()
                 .map_err(|_| ExceptionalHalt::VeryLargeNumber)?;
             let return_data_size: usize = return_data_size
@@ -145,6 +151,13 @@ impl<'a> VM<'a> {
             let args_size = args_size
                 .try_into()
                 .map_err(|_err| ExceptionalHalt::VeryLargeNumber)?;
+            let args_start_offset = match args_start_offset.try_into() {
+                Ok(x) => x,
+                Err(_) => usize::MAX,
+            };
+            let return_data_start_offset = return_data_start_offset
+                .try_into()
+                .map_err(|_err| ExceptionalHalt::VeryLargeNumber)?;
             let return_data_size = return_data_size
                 .try_into()
                 .map_err(|_err| ExceptionalHalt::VeryLargeNumber)?;
@@ -227,16 +240,18 @@ impl<'a> VM<'a> {
             return Ok(OpcodeResult::Halt);
         }
 
+        let offset: usize = match offset.try_into() {
+            Ok(x) => x,
+            Err(_) => usize::MAX,
+        };
+
         let new_memory_size = calculate_memory_size(offset, size)?;
         let current_memory_size = current_call_frame.memory.len();
 
         current_call_frame
             .increase_consumed_gas(gas_cost::exit_opcode(new_memory_size, current_memory_size)?)?;
 
-        current_call_frame.output =
-            memory::load_range(&mut current_call_frame.memory, offset, size)?
-                .to_vec()
-                .into();
+        current_call_frame.output = current_call_frame.memory.load_range(offset, size)?.into();
 
         Ok(OpcodeResult::Halt)
     }
@@ -263,7 +278,13 @@ impl<'a> VM<'a> {
                 return_data_size,
             ] = *current_call_frame.stack.pop()?;
             let address = word_to_address(address);
+            let args_start_offset = args_start_offset
+                .try_into()
+                .map_err(|_err| ExceptionalHalt::VeryLargeNumber)?;
             let args_size = args_size
+                .try_into()
+                .map_err(|_err| ExceptionalHalt::VeryLargeNumber)?;
+            let return_data_start_offset = return_data_start_offset
                 .try_into()
                 .map_err(|_err| ExceptionalHalt::VeryLargeNumber)?;
             let return_data_size = return_data_size
@@ -358,7 +379,13 @@ impl<'a> VM<'a> {
                 return_data_size,
             ] = *current_call_frame.stack.pop()?;
             let address = word_to_address(address);
+            let args_start_offset = args_start_offset
+                .try_into()
+                .map_err(|_err| ExceptionalHalt::VeryLargeNumber)?;
             let args_size = args_size
+                .try_into()
+                .map_err(|_err| ExceptionalHalt::VeryLargeNumber)?;
+            let return_data_start_offset = return_data_start_offset
                 .try_into()
                 .map_err(|_err| ExceptionalHalt::VeryLargeNumber)?;
             let return_data_size = return_data_size
@@ -442,6 +469,10 @@ impl<'a> VM<'a> {
         let code_size_in_memory: usize = code_size_in_memory
             .try_into()
             .map_err(|_err| ExceptionalHalt::VeryLargeNumber)?;
+        let code_offset_in_memory: usize = match code_offset_in_memory.try_into() {
+            Ok(x) => x,
+            Err(_) => usize::MAX,
+        };
 
         let new_size = calculate_memory_size(code_offset_in_memory, code_size_in_memory)?;
 
@@ -473,6 +504,10 @@ impl<'a> VM<'a> {
         let code_size_in_memory: usize = code_size_in_memory
             .try_into()
             .map_err(|_err| ExceptionalHalt::VeryLargeNumber)?;
+        let code_offset_in_memory: usize = match code_offset_in_memory.try_into() {
+            Ok(x) => x,
+            Err(_) => usize::MAX,
+        };
 
         let new_size = calculate_memory_size(code_offset_in_memory, code_size_in_memory)?;
 
@@ -500,6 +535,12 @@ impl<'a> VM<'a> {
         let current_call_frame = self.current_call_frame_mut()?;
 
         let [offset, size] = *current_call_frame.stack.pop()?;
+
+        let offset = match offset.try_into() {
+            Ok(x) => x,
+            Err(_) => usize::MAX,
+        };
+
         let size = size
             .try_into()
             .map_err(|_| ExceptionalHalt::VeryLargeNumber)?;
@@ -510,10 +551,7 @@ impl<'a> VM<'a> {
         current_call_frame
             .increase_consumed_gas(gas_cost::exit_opcode(new_memory_size, current_memory_size)?)?;
 
-        current_call_frame.output =
-            memory::load_range(&mut current_call_frame.memory, offset, size)?
-                .to_vec()
-                .into();
+        current_call_frame.output = current_call_frame.memory.load_range(offset, size)?.into();
 
         Err(VMError::RevertOpcode)
     }
@@ -589,7 +627,7 @@ impl<'a> VM<'a> {
     pub fn generic_create(
         &mut self,
         value: U256,
-        code_offset_in_memory: U256,
+        code_offset_in_memory: usize,
         code_size_in_memory: usize,
         salt: Option<U256>,
     ) -> Result<OpcodeResult, VMError> {
@@ -614,12 +652,9 @@ impl<'a> VM<'a> {
 
         // Load code from memory
         let code = Bytes::from(
-            memory::load_range(
-                &mut self.current_call_frame_mut()?.memory,
-                code_offset_in_memory,
-                code_size_in_memory,
-            )?
-            .to_vec(),
+            self.current_call_frame_mut()?
+                .memory
+                .load_range(code_offset_in_memory, code_size_in_memory)?,
         );
 
         // Get account info of deployer
@@ -683,6 +718,8 @@ impl<'a> VM<'a> {
         let mut stack = self.stack_pool.pop().unwrap_or_default();
         stack.clear();
 
+        let next_memory = self.current_call_frame()?.memory.next_memory();
+
         let new_call_frame = CallFrame::new(
             deployer,
             new_address,
@@ -695,9 +732,10 @@ impl<'a> VM<'a> {
             new_depth,
             true,
             true,
-            U256::zero(),
+            0,
             0,
             stack,
+            next_memory,
         );
         self.call_frames.push(new_call_frame);
 
@@ -716,6 +754,10 @@ impl<'a> VM<'a> {
     /// This (should) be the only function where gas is used as a
     /// U256. This is because we have to use the values that are
     /// pushed to the stack.
+    ///
+    // Force inline, due to lot of arguments, inlining must be forced, and it is actually beneficial
+    // because passing so much data is costly. Verified with samply.
+    #[inline(always)]
     pub fn generic_call(
         &mut self,
         gas_limit: u64,
@@ -726,7 +768,7 @@ impl<'a> VM<'a> {
         should_transfer_value: bool,
         is_static: bool,
         calldata: Bytes,
-        ret_offset: U256,
+        ret_offset: usize,
         ret_size: usize,
         bytecode: Bytes,
         is_delegation_7702: bool,
@@ -755,6 +797,8 @@ impl<'a> VM<'a> {
         let mut stack = self.stack_pool.pop().unwrap_or_default();
         stack.clear();
 
+        let next_memory = self.current_call_frame_mut()?.memory.next_memory();
+
         let new_call_frame = CallFrame::new(
             msg_sender,
             to,
@@ -770,6 +814,7 @@ impl<'a> VM<'a> {
             ret_offset,
             ret_size,
             stack,
+            next_memory,
         );
         self.call_frames.push(new_call_frame);
 
@@ -828,8 +873,11 @@ impl<'a> VM<'a> {
             gas_limit,
             ret_offset,
             ret_size,
+            memory: old_callframe_memory,
             ..
         } = executed_call_frame;
+
+        old_callframe_memory.clean_from_base();
 
         let parent_call_frame = self.current_call_frame_mut()?;
 
@@ -843,12 +891,25 @@ impl<'a> VM<'a> {
             .ok_or(InternalError::Overflow)?;
 
         // Store return data of sub-context
-        memory::try_store_range(
-            &mut parent_call_frame.memory,
+        if ret_size > 0 && ctx_result.output.len() < ret_size {
+            parent_call_frame.memory.resize(
+                ret_offset
+                    .checked_add(ret_size)
+                    .ok_or(ExceptionalHalt::OutOfBounds)?,
+            )?;
+        }
+        parent_call_frame.memory.store_data(
             ret_offset,
-            ret_size,
-            &ctx_result.output,
+            if ctx_result.output.len() >= ret_size {
+                ctx_result
+                    .output
+                    .get(..ret_size)
+                    .ok_or(ExceptionalHalt::OutOfBounds)?
+            } else {
+                &ctx_result.output
+            },
         )?;
+
         parent_call_frame.sub_return_data = ctx_result.output.clone();
 
         // What to do, depending on TxResult
@@ -880,8 +941,12 @@ impl<'a> VM<'a> {
             gas_limit,
             to,
             call_frame_backup,
+            memory: old_callframe_memory,
             ..
         } = executed_call_frame;
+
+        old_callframe_memory.clean_from_base();
+
         let parent_call_frame = self.current_call_frame_mut()?;
 
         // Return unused gas
@@ -921,9 +986,9 @@ impl<'a> VM<'a> {
     /// Obtains the values needed for CALL, CALLCODE, DELEGATECALL and STATICCALL opcodes to calculate total gas cost
     fn get_call_gas_params(
         &mut self,
-        args_start_offset: U256,
+        args_start_offset: usize,
         args_size: usize,
-        return_data_start_offset: U256,
+        return_data_start_offset: usize,
         return_data_size: usize,
         eip7702_gas_consumed: u64,
         address: Address,
@@ -952,9 +1017,11 @@ impl<'a> VM<'a> {
         ))
     }
 
-    fn get_calldata(&mut self, offset: U256, size: usize) -> Result<Bytes, VMError> {
+    fn get_calldata(&mut self, offset: usize, size: usize) -> Result<Bytes, VMError> {
         Ok(Bytes::from(
-            memory::load_range(&mut self.current_call_frame_mut()?.memory, offset, size)?.to_vec(),
+            self.current_call_frame_mut()?
+                .memory
+                .load_range(offset, size)?,
         ))
     }